tests/wpt/web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.get.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.imageData.get.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.get.invalid</h1>
<p class="desc">Verify getImageData() behavior in invalid cases.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Verify getImageData() behavior in invalid cases.");
_addTest(function(canvas, ctx) {

imageData = ctx.getImageData(0,0,2,2);
var testValues = [NaN, true, false, "\"garbage\"", "-1",
                  "0", "1", "2", Infinity, -Infinity,
                  -5, -0.5, 0, 0.5, 5,
                  5.4, 255, 256, null, undefined];
var testResults = [0, 1, 0, 0, 0,
                   0, 1, 2, 255, 0,
                   0, 0, 0, 0, 5,
                   5, 255, 255, 0, 0];
for (var i = 0; i < testValues.length; i++) {
    imageData.data[0] = testValues[i];
    _assert(imageData.data[0] == testResults[i], "imageData.data[\""+(0)+"\"] == testResults[\""+(i)+"\"]");
}
imageData.data['foo']='garbage';
_assert(imageData.data['foo'] == 'garbage', "imageData.data['foo'] == 'garbage'");
imageData.data[-1]='garbage';
_assert(imageData.data[-1] == undefined, "imageData.data[-1] == undefined");
imageData.data[17]='garbage';
_assert(imageData.data[17] == undefined, "imageData.data[\""+(17)+"\"] == undefined");


});
</script>
